Fix an important bug
This revision fixes an important bug and makes several minor changes.
The bug was that the current time was increased before the propagation
step was made.  Because of this, there was absolutely no change in the
wave function between the beginning of a simulation and the end of the
first time step.

Fixing the bug, I realized that the implementation of a fixed time
step was extremely confusing and not quite right.